from one computer to another

+++ RayMarch/src/Light.java
@@ -1,9 +1,9 @@
 public class Light extends Vertex3<Double> {
-    public final byte r;
-    public final byte g;
-    public final byte b;
+    public final double r;
+    public final double g;
+    public final double b;
     public final double intensity;
-    public Light(double x, double y, double z, byte r, byte g, byte b, double intensity) {
+    public Light(double x, double y, double z, double r, double g, double b, double intensity) {
         super(x, y, z);
         this.r = r;
         this.g = g;
